pyproject.toml
[tool.poetry]
name = "aineko-dream"
version = "0.0.0"
description = "Dreaming of Aineko apps..."
authors = ["Aineko Engineering <engineering@aineko.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
aineko = "0.2.3"
openai = "^1.2.4"
uvicorn = "^0.24.0.post1"
pygithub = "^2.1.1"
python-dotenv = "^1.0.0"
fastapi = "^0.104.1"
bandit = "^1.7.5"
aiokafka = "^0.8.1"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.1"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
aineko = "aineko.__main__:_cli"
